reference/transcription-mcp-v2/pyproject.toml
[project]
name = "transcription-mcp"
version = "2.0.0"
description = "MCP server for bilingual transcription with WhisperX"
authors = [{name = "Frans Vermaak", email = "frans@larcai.com"}]
requires-python = ">=3.10"
readme = "README.md"

dependencies = [
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "whisperx>=3.1.0",
    "python-docx>=1.0.0",
    "aiofiles>=23.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/transcription_mcp"]
